commit 2

+++ BasicSalesSystem.Web/Custom/Resources/PagedListResource.cs
@@ -5,12 +5,12 @@
     public sealed class PagedListResource
     {
         public IEnumerable<object> ItemsList { get; set; }
-        public int TotalItemCount { get; set; }
+        public long TotalItemCount { get; set; }
 
         public PagedListResource()
         {
         }
-        public PagedListResource(IEnumerable<object> itemsList, int totalItemCount)
+        public PagedListResource(IEnumerable<object> itemsList, long totalItemCount)
             : this()
         {
             ItemsList = itemsList;
